requirements.txt
pypdf
streamlit
dotenv
langchain
langchain-google-genai
langchain-community
google-generativeai
faiss-cpu
pip install PyPDF2